Add tests
I also refactored the test fixtures so that they are shared across
tests.
